requirements.txt
functions-framework==3.5.0
flask==3.0.3
google-cloud-error-reporting==1.9.1
MarkupSafe==2.1.3
requests
google-cloud-secret-manager
google-cloud-resource-manager
